Updated testing for detail page
