+++ client/__mocks__/data.json
@@ -1,0 +1,54 @@
+{
+    "mockData":{
+        "detail":{
+            "data":{
+                "detail":{  
+                    "overview": "test_overview",
+                    "genre": "test_genre",
+                    "tagline": "test_tagline",
+                    "hours": "test_hours",
+                    "minutes": "test_minutes",
+                    "video": "test_video"
+                },
+                "movieList": [ 
+                    {
+                        "id": "test_id_1",
+                        "title": "test_title_1",
+                        "poster": "test_poster_1",
+                        "date": "test_release_date_1",
+                        "voteAverage": "test_vote_average_1" 
+                    },
+                    {
+                        "id": "test_id_2",
+                        "title": "test_title_2",
+                        "poster": "test_poster_2",
+                        "date": "test_release_date_2",
+                        "voteAverage": "test_vote_average_2" 
+                    }
+                ],
+                "reviews":[
+                    {
+                        "author": "test_author_1",
+                        "content": "test_content_1"
+                    },
+                    {
+                        "author": "test_author_2",
+                        "content": "test_content_2"
+                    }
+                ]
+            }
+        },
+        "error":{
+            "response": {
+                "data":{
+                    "errorMsg":"test_api_error"
+                }
+            }
+        },
+        "backendError":{
+            "response":{
+                "statusText": "test_backend_error"
+            }
+        }
+    }
+}
