+++ client/src/containers/DetailPage/Detail.js
@@ -40,7 +40,7 @@ class Detail extends Component{
             })
         }).catch((error)=>{
             this.setState({
-                errorMsg: error.response.data.errorMsg,
+                errorMsg: error.response.data? error.response.data.errorMsg: error.response.statusText,
                 loading: false
             })
         });
