+++ client/src/containers/Home/Home.js
@@ -54,7 +54,7 @@ class Home extends Component {
             })
         }).catch((error)=>{
             this.setState({
-                errorMsg: error.response.data.errorMsg,
+                errorMsg: error.response.data.errorMsg? error.response.data.errorMsg: error.response.statusText,
                 loading: false
             })
         })
